app/requirements.txt
biopython==1.73
pubmed-lookup==0.2.1
google-api-python-client==1.7.8
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
pygsheets==2.0.1
pandas==0.24.2